Add not yet Working IACTFile and CorsikaEvent, define API in tests
